Update text types to allow CJK analyzers
The CJK analyzers don't like it when the number of characters
in a single-tokenized query changes. This was true for the
old way we were dealing with left-anchored ("starts with") and
fully_anchored ("exactish") fieldTypes.

The updates the `.jar` files to use the latest available for
both the `icu` jars and the local, custom `lucene-umich-solr-filters`,
similar to what the umich catalog is using. This should give us
both no errors for CJK items and better anchored matching
all around.

TODO: Update the whole solr schema using knowledge gained from
working on the UMich catalog.