Remove dead logging code and skip logging specs
rspec futzes with stderr/stdout when running for purposes
of capturing with `output`.

> Note: to_stdout and to_stderr work by temporarily replacing $stdout or $stderr,
> so they're not able to intercept stream output that explicitly uses STDOUT/STDERR
> or that uses a reference to $stdout/$stderr that was stored before the matcher was used.

(from https://www.rubydoc.info/gems/rspec-expectations/RSpec%2FMatchers:output)

Because we create the logger object long before the tests are run, and because
the logger is a class singleton (i.e., we can't just make a new one with the
new, rspec-aliased $stderr), our `output` matchers aren't going to work.

We're not the only ones, of course:
https://github.com/reidmorrison/rails_semantic_logger/issues/139

So...skipping them for now. Real solution is probably to have LoggerFactory
actually be a factory and produce a logger which can be assigned to
the Services object, or maybe use `include SemanticLogger::Loggable`
everywhere.