Change CICTL structure so shared options display with help
Thor `class_option` functionality is passed along to subcommands
mounted with the `subcommand` directive, but the help text
for those subcommands does not correctly display the options
defined with `class_option`.

This adds a `BaseCommand` class that defines the top level
options (--log and --verbose), and the subcommands
inherit from it rather than from `Thor`. Also renames
the class that mounts the subcommands from `Command` to
`Commands` to try to make the naming slightly less
janky.

Also strengthen delete file filter to remove non-numeric, in
addition to empty, lines.